feat: extract domain layer with repository interfaces
- core:domain 모듈 추가
  - Repository 인터페이스를 core:data에서 core:domain으로 이동
- core:data 모듈이 core:domain을 참조하도록 의존성 방향 변경
- presentation 모듈이 core:domain을 참조하도록 수정
- presentation 모듈의 build.gradle.kts에서 불필요한 core:data 의존성 제거
