+++ core/domain/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    alias(libs.plugins.aiku.android.library)
+}
+
+android {
+    namespace = "com.hyunjung.aiku.core.domain"
+}
+
+dependencies {
+    api(projects.core.common)
+    api(projects.core.model)
+    api(libs.kotlinx.coroutines.core)
+}
