+++ core/domain/src/main/kotlin/com/hyunjung/aiku/core/domain/repository/GroupRepository.kt
@@ -1,4 +1,4 @@
-package com.hyunjung.aiku.core.data.repository
+package com.hyunjung.aiku.core.domain.repository
 
 import com.hyunjung.aiku.core.model.GroupDetail
 import com.hyunjung.aiku.core.model.GroupOverview
