+++ settings.gradle.kts
@@ -33,6 +33,7 @@ include(":core")
 include(":core:common")
 include(":core:data")
 include(":core:designsystem")
+include(":core:domain")
 include(":core:navigation")
 include(":core:network")
 include(":core:model")
